requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.8
django-celery-beat==2.6.0
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
kombu==5.4.0
Markdown==3.6
prompt_toolkit==3.0.47
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
